Fix ninja target: check-upgrade-activated
After merging of !1891 & !1894, we need to update the cmake system to
properly update the ninja target `check-upgrade-activated`.

It turns out that with uprade11 activated, a couple of functional tests no
longer fully work due to changed node error messages and/or changes in
consensus after upgrade 11.

For the tests in question, we force upgrade11 to be deactivated for the
tests so that the tests continue to work (they are perfectly ok as
regression tests for pre-upgrade 11 parts of the chain).