Prevent UB in index/base.* my removing pure virtual GetName() function
This is a partial backport of core commit:
https://github.com/bitcoin/bitcoin/pull/25971/commits/200d84d5681918523d982b9ddf60d1127edcb44

The reason we need this commit is because as implemented, the
`BaseIndex` class was potentially accessing the pure virtual function
`GetName()` from its destructor `~BaseIndex()`, which could lead to UB and/or
crashes.

We instead make `GetName()` be non-virtual and we have it refer to an
`m_name` variable whose lifetime is as long as the `BaseIndex` itself.

Co-authored-by: stickies-v <stickies-v@protonmail.com>