Update static seeds in preparation for v28 release
Used the procedure described in contrib/seeds/README.md to update the
static seeds for mainnet. Note that for some reason the static seeds we
had for testnet3, testnet4, and chipnet were not synched to what is in
the header file in chainparamsseeds.h and this commit also synchs them
up.

I also manually added blackie.c3-soft.com for test3, test4, and chipnet.