Update checkpoints for v28.0.0 release
Summary
---

The checkpoints for mainnet, testnet3, testnet4, and chipnet are updated
by this commit.  They are all >= 10 blocks in the past for mainnet,
>=2000 blocks in the past for the testnet(s) except for chipnet where we
recently had a network upgrade so the checkpoint is a few hundred blocks
after the network upgrade, which itself was several hundred blocks ago.

Test Plan
---

Use a block explorer e.g. https://bch.loping.net to verify the heights
and hash match for each of the 4 entries added for each respective chain.