refactor: Move class `BlockValidationOptions` to consensus/validation.h
We put class `BlockValidationOptions` into a the much smaller & lighter
file: `<consensus/validation.h>`. Previously it lived in the very large
top-level `<validation.h>`.  We need to do this so we can re-use this
class without needing to include all of the monstrous `<validation.h>`.

This will be needed by a subsequent commit.

Also in this commit: Slight refactor/cleanup of some stuff in
`consensus/validation.h`