[net] Parallel compact block downloads
Co-authored-by: Gregory Sanders <gsanders87@gmail.com>

This is a partial backport of
[Core#27752](https://github.com/bitcoin/bitcoin/pull/27752).  It
improves performance and lowers latency for newly mined blocks and their
propagation, by allowing for up to 3 parallel compactblock requests to
be enqueued at once.  For more information see the Core link.

This has been heavily adapted and refactored by me to remove some warts
from the original commit(s).