net: Remove extraneous/bad log line from net_processing.cpp
This log line appeared in Bitcoin-ABC from a mis-applied backport of Core.
The actual log line is already in `FinalizeNode()`, but the backport in
ABC put it again in `PushNodeVersion` where it makes no sense to be.

Bad ABC backport: https://reviews.bitcoinabc.org/D3284

This commit just deletes the line since it makes no sense to say
"cleared nodestate" in PushNodeVersion() (instead it should appear in
FinalizeNode, where it does indeed exist).