net: Avoid excess CPU usage by `msghand` thread; fix "early wake up" logic
Fixes issue #535. The `msghand` thread was waking up excessively even
when there was no work to be done. This is described in detail in #535.

This commit fixes the situation by having it wake up early *only* if it
knows that there will be work to be done in the near future.

This avoids wasteful excessing wake-ups and wasteful CPU consumption by
this thread.