trivial: Remove unused/aborting function from class CScript
The function operator<<(CScript) was implemented but would assert(false)
if called. This is dangerous since inadvertently it could be called by
new code and the code would compile fine, only to abort() at runtime.

Having such intentional assert(false) in the codebase is bad especially
when a compile-time check could catch this type of programmer misuse.

Instead, since we explicitly want to forbid operator<< on CScripts, we
should just mark it as deleted and document why. This way any
inadvertent use can be caught at compile-time, rather than at runtime.

This commit does just that.