Update DisconnectedBlockTransactions::addForBlock to use CHECK_NONFATAL rather than assert()
This is a code quality improvement. The boost unit tests should not be
asserting and aborting since this screws up the logging of the app. It's
better for the unit tests that test the
DisconnectedBlockTransactions::addForBlock to instead simply throw if
`checkSanity=true`, so boost can detect the error and log properly.