mempool: Fix potential for bug in mempool removeForBlock()
The mempool could end up in an inconsistent state in
`CTxMemPool::removeForBlock()`. This is due to the fact that it was
using a temporary `DisconnectedBlockTransactions` pool to topologically
sort the block transactions it was receiving from the caller. However,
this temporary pool's `addForBlock` function might drop txns if the
block size exceeds the limits of the pool (currently ~640MB dynamic
size, about 300MB block size).

The dropping of these transactions could lead to the mempool not
realizing this happened, and conflicts in the block not being detected
properly versus the mempool.

This potential bug cannot be triggered on the current chain but might be
able to be triggered someday if blocks grow beyond 300 MB.

The reason for the bug is that the disconnect pool is doing double-duty.
Sometimes it's used as a disconnect pool, and sometimes it's used as a
utility class for topological sorting of transactions.

This commit does the following:

- separates out addForBlock into two functions: addNoLimit, and
  addForBlock.
  - addForBlock behaves exactly as before
  - addNoLimit is like addForBlock but it doesn't enforce limits. This
    is the one to call if using the pool as a utility topo sorter.
- Changes the two call sites where the disconnectpool was being used as
  a glorified topo sorted, to instead use addNoLimit -- since limiting
  size at those call sites buys use nothing (the txns are in memory
  anyway!!) and the temporary pool will go away/be emptied immediately
  anyway.
- Adds unit tests to test the API changes to
  DisconnectedBlockTransactions.